+++ lib/gltfparse.js
@@ -1,5 +1,7 @@
 import { GLTFLoader } from '../node_modules/three/examples/jsm/loaders/GLTFLoader.js';
 import { defaultValues } from './schema.js';
+import common from './glsl/common.glsl.js';
+import mapFragment from './glsl/map_fragment.glsl.js';
 import {
     Group,
     Vector3
@@ -86,89 +88,12 @@ const initPromise = (async function init() {
         }
     });
 
+    /* Tweak the material to enable the map to behave as a HSL indexed colour map rather than as an rgb map */
     material.onBeforeCompile = function (shader) {
         console.log({uniforms: shader.uniforms});
-
         shader.uniforms.xravatar_palette = palette;
-
-        /*
-        Tweak the map to use the map as an index rather than as an image
-
-        Index = floor(r/16);
-        Brightness = r%16;
-
-        Use this to get the colour in HSL format from the array of 16 colours in the uniforms
-
-        put into hsl2rgb to get final colour.
-        */
-        shader.fragmentShader = shader.fragmentShader.replace('#include <map_fragment>', `
-        #ifdef USE_MAP
-            vec4 texelColor = texture2D( map, vUv );
-            texelColor = mapTexelToLinear( texelColor );
-            int index = int(16.0 * texelColor.r);
-            float bshift = 0.5 * 16.0 * (mod(texelColor.r, 0.0625) - 0.03125);
-            vec3 hsl;
-            ${
-                shader.uniforms.xravatar_palette.value.map((o,i) => `if ( ${i} <= index ) hsl = xravatar_palette[ ${i} ];`).join('\n')
-            }
-            vec3 rgb = hsl2rgb(hsl.x/360.0, hsl.y, hsl.z + bshift);
-            diffuseColor *= vec4(rgb,1.0);
-        #endif
-        `)
-
-
-        /* Include some helper functions, and our uniform after <common> */
-        shader.fragmentShader = shader.fragmentShader.replace('#include <common>', `
-        uniform vec3 xravatar_palette[16];
-
-        #include <common>
-
-        // From https://github.com/Jam3/glsl-hsl2rgb/blob/master/index.glsl
-		// MIT License (MIT) Copyright (c) 2015 Jam3
-		
-		float hue2rgb(float f1, float f2, float hue) {
-			if (hue < 0.0)
-				hue += 1.0;
-			else if (hue > 1.0)
-				hue -= 1.0;
-			float res;
-			if ((6.0 * hue) < 1.0)
-				res = f1 + (f2 - f1) * 6.0 * hue;
-			else if ((2.0 * hue) < 1.0)
-				res = f2;
-			else if ((3.0 * hue) < 2.0)
-				res = f1 + (f2 - f1) * ((2.0 / 3.0) - hue) * 6.0;
-			else
-				res = f1;
-			return res;
-		}
-		
-		vec3 hsl2rgb(vec3 hsl) {
-			vec3 rgb;
-			
-			if (hsl.y == 0.0) {
-				rgb = vec3(hsl.z); // Luminance
-			} else {
-				float f2;
-				
-				if (hsl.z < 0.5)
-					f2 = hsl.z * (1.0 + hsl.y);
-				else
-					f2 = hsl.z + hsl.y - hsl.y * hsl.z;
-					
-				float f1 = 2.0 * hsl.z - f2;
-				
-				rgb.r = hue2rgb(f1, f2, hsl.x + (1.0/3.0));
-				rgb.g = hue2rgb(f1, f2, hsl.x);
-				rgb.b = hue2rgb(f1, f2, hsl.x - (1.0/3.0));
-			}   
-			return rgb;
-		}
-		
-		vec3 hsl2rgb(float h, float s, float l) {
-			return hsl2rgb(vec3(h, s, l));
-		}
-        `);
+        shader.fragmentShader = shader.fragmentShader.replace('#include <map_fragment>', mapFragment(shader))
+        shader.fragmentShader = shader.fragmentShader.replace('#include <common>', common(shader));
     };
 
     toBeRemoved.forEach(o => o.parent.remove(o));
